requirements.txt
fastapi
httpx
uvicorn
google-generativeai
